+++ tests/north_project/sql/1.0/1.0-author-2-dml.sql
@@ -9,5 +9,7 @@ INSERT INTO auth_permission(codename, name, content_type_id) VALUES('add_author'
 INSERT INTO auth_permission(codename, name, content_type_id) VALUES('change_author', 'Can change author', (SELECT id FROM django_content_type WHERE app_label = 'north_app' AND model = 'author'));
 INSERT INTO auth_permission(codename, name, content_type_id) VALUES('delete_author', 'Can delete author', (SELECT id FROM django_content_type WHERE app_label = 'north_app' AND model = 'author'));
 
+-- Django >= 2.1
+INSERT INTO auth_permission(codename, name, content_type_id) VALUES('view_author', 'Can view author', (SELECT id FROM django_content_type WHERE app_label = 'north_app' AND model = 'author'));
 
 COMMIT;