+++ tests/north_project/sql/1.0/1.0-book-2-dml.sql
@@ -9,5 +9,7 @@ INSERT INTO auth_permission(codename, name, content_type_id) VALUES('add_book', 
 INSERT INTO auth_permission(codename, name, content_type_id) VALUES('change_book', 'Can change book', (SELECT id FROM django_content_type WHERE app_label = 'north_app' AND model = 'book'));
 INSERT INTO auth_permission(codename, name, content_type_id) VALUES('delete_book', 'Can delete book', (SELECT id FROM django_content_type WHERE app_label = 'north_app' AND model = 'book'));
 
+-- Django >= 2.1
+INSERT INTO auth_permission(codename, name, content_type_id) VALUES('view_book', 'Can view book', (SELECT id FROM django_content_type WHERE app_label = 'north_app' AND model = 'book'));
 
 COMMIT;